prevent upgrades from 1.1 (force uninstall)

+++ spk/gateone/src/installer.sh
@@ -86,6 +86,14 @@ preupgrade ()
 {
     # Stop the package
     ${SSS} stop > /dev/null
+    
+    # version 1.2 is not backwards compatible with 1.1
+    if [ `echo ${SYNOPKG_OLD_PKGVER} | sed -r -e 's/^([0-9]+)\.[0-9]+-[0-9]+$/\1/'` -lt 2 ] && [ `echo ${SYNOPKG_OLD_PKGVER} | sed -r -e 's/^[0-9]+\.([0-9]+)-[0-9]+$/\1/'` -lt 2 ]; then
+        echo "Please uninstall previous version, no update possible" 
+        echo "Remember to save your server.conf file before uninstalling"
+        echo "You will need to manually port old configuration settings to the new configuration files in 1.2"
+        exit 1
+    fi
 
     # Save some stuff
     rm -fr ${TMP_DIR}/${PACKAGE}
